requirements.txt
flask==2.3.2
requests==2.31.0
flask-sqlalchemy==3.0.3
plotly==5.15.0
python-dotenv==1.0.0
werkzeug
pycoingecko
gunicorn
